+++ pizzeria/src/main/java/it/pizzeria/pizzeria/repository/OffertaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.pizzeria.pizzeria.model.Offerta;
 
-public interface OffertaRepository extends JpaRepository <Offerta, Integer> {
+public interface OffertaRepository extends JpaRepository <Offerta, Long> {
 
 
 }
